Add export utilities and feature flags
